requirements.txt
numpy
imageio
Pillow